merge Session and SessionImpl into a single class
Based on a discussion with gpolitis in #jitsi on 2014-02-03, we both
thought it made a lot of sense to get rid of the *Impl classes and
corresponding abstract classes since each abstract class only had a single
subclass, the *Impl version.

Merging these also makes it much easier to write tests for all of the
classes that are split between an abstract and *Impl.