document getFallbackMessage() for QTR query, and include default
It has become standard for the OTR query message to include a message to a
human who is receiving the OTR query on a client that does not support OTR.
This adds a default message, in case getFallbackMessage() only returns a null
or "".  It also serves as an example for what people should write in
getFallbackMessage() for their app.