Provide a checkstyle configuration
Provides a checkstyle configuration closely aligned with the android
code style guidelines (which are quite vague) + a lot of additional
checks which are underspecified in the android style guide. Checking is
integrated in the maven build system by using the maven reporting
features.

The checkstyle report will be generated when calling `mvn site`. It is
available as a HTML page under `target/site/index.html`

* codecheck/checkstyle.xml: new file, main checkstyle configuration
* codecheck/checkstyle-suppressions.xml:
  new file: suppressions file to
  prevent certain checks from being applied to specific files. Currently
  used to disable javadoc checks on unit tests
* pom.xml: enabled reporting and added checkstyle as a report