Remove the maven jitsy repo from README.markdown
Since this fork does not use the jitsi repository for deploying releases
of otr4j, it should not be mentioned in the README file. Otherwise,
users might get confused and end up with binary version of the library
from the jitsi version and not the one they wanted to have.