SMPTest: include unicode tests for both question/answer and password
Since SMP verification is based on user input, it should handle all manner
of unicode that anything might throw at it.  It should also handle badly
formed strings.  The downside is that this test suite takes 5 minutes to
run because of the whole OTR and SMP setup for each test session.

Some of these are failing, it seems that the question handling does not
properly sanitize the input, or if it is meant to accept special chars,
then it is not handling them correctly.

refs https://github.com/otr4j/otr4j/issues/30
closes https://github.com/otr4j/otr4j/pull/16