automatically check available PGP signatures on jars downloaded by maven
HTTPS is not enough to make sure that the jar is legit.  HTTPS protects the
network traffic.  The PGP signature helps verify that the file up on the
maven repository is the same as the file that the original.  But it seems
like the PGP implementation in maven repos leaves a lot to be desired
since there is no canonical keyring to test against, like for example with
Debian repositories.  But this is better than nothing, and works
automatically.