parse DataMessages as byte[] to fix Unicode issues
The original code was first parsing the message body into a String, then
basing the TLV parsing off of the length of that String.  That works for
ASCII, where one char is one byte, but totally fails for non-ASCII, since
basically all non-latin chars are more than one byte.

This also has the advantage of removing the trailing null char from the end
of the message body itself.

closes https://github.com/otr4j/otr4j/issues/46
closes https://github.com/otr4j/otr4j/pull/50