Use utility class for safely handling OtrEngineListener calling.
Provide a utility that safely handles calling listeners of type
OtrEngineListener. Listeners will be handled in sequence through
iteration. Any thrown RuntimeException exceptions are caught and logged
in order to inform the user. Runtime exceptions will not escape the
utility thus ensuring that the OTR program logic will not be interrupted
unexpectedly.