Resolved remaining instances of readInt and documented interpretation issues.
Due to Java's signed interpretation of int values, there are instances
where integer values read from OtrInputStream are interpreted as signed.
In some cases this can lead to runtime exceptions. These cases are
evaluated and in some cases support is limited to what is supported by
Java. This is a limitation of otr4j, as OTR specifies support for full 4
byte unsigned integer values.

In other cases, the full int value is used, as these cases use the
integer value as identifier. These cases need not interpret the value as
an (signed/unsigned) integer value.