Replaced obsolete Hashtable with (synchronized) HashMap.
To get rid of obsolete collection type Hashtable, we replace it with
HashMap. Because otr4j inherently has some concurrency aspects, the
HashMap instance is synchronized.

There is still an open TODO for considering whether or not to do
'contains' check + 'put' operation on sessions map inside a
synchronized-block. In theory this should absolutely be synchronized,
however, the chance that the same sessionID is inserted twice
concurrently is very small, so I'll leave this as a separate
consideration for now.