Removed TODO: should be fixed in some other way.
Although it would be good to store these parameters separately, such
that it is *only* possible to access the fields through accessors. Right
now it looks like this is only really possible with the introduction of
the State pattern that reflects the state machine of OTR. That would
solve most of the issues that we tried to tackle by putting the data in
a separate container.