Decided not to throw an exception as getFingerprint() only supports OtrEngineHost calls.
There is only a risk of calling OtrEngineHost with null fingerprint when
the fingerprint cannot be acquired. Furthermore, as the required
instance is already needed before this stage, the chance for this error
case is very small. Accept that simply logging the exception if it
occurs is sufficient.
This also means that there is a chance that OtrEngineHost is called with
a null fingerprint argument.