Removed TODO: obsolete. Decorator is not the right solution.
In general the idea was okay, but there is a caveat: The decorator works
in principle but we are looking for additional guarantees, i.e. the
guarantee that the message handling process will not be (unnecessarily)
interrupted.
We currently solve this with utility methods that wrap OtrEngineHost
calls in a try-catch block and catch all harmful exceptions. A decorator
would work for wrapping the OtrEngineHost instance and ensure safe
execution, however that means that we do not leverage the type system to
guarantee safety.
We could introduce a new "SafeOtrEngineHost" interface that requires
that no checked exceptions are thrown (enforced by syntax) and no
runtime exceptions are thrown (enforced by contract/description).
However, I think that the utility methods solve enough of this problem
that the SafeOtrEngineHost interface is not necessary.
Furthermore, not all methods can be called safely, as we depend on the
correct, full execution of these methods in order to progress in our
message handling.