Synchronized session creation.
Synchronized full session retrieval code. In case session does not yet
exist, this is detected and resolved within the synchronized block such
that there is no possibility of a concurrent session *also* starting to
create a new session.

As documented, we do not distinguish between existing entry but with
null session and non-existing entry. This is completely irrelevant as in
both cases we cannot acquire a valid session.