Fixed null handling (mostly). Remove null string from unicode test strings.
Nulls have a different effect in different cases. In case of parsing an
OTR encoded message where the plain text message part contains NULL
characters, it is not possible to distinguish between invalid NULL and
valid NULL separating message from TLV. This was solved in an earlier
commit by doing a sanity check on large integer values to prevent
unintentionally interpreting ASCII chars as part of integer value.

Now two tests are added that verify behavior with serializing strings
that contain NULL characters. For plain text messages, no special
treatment is in place. For byte-conversions, we replace NULL by ? to
ensure that we do not accidentally produce a badly constructed OTR
encoded message.

Now that we handle both receiving and sending case for NULL chars, the
string containing NULL chars can be removed from the list of unicode
test strings. It is not possible to make it part of the regular test
plan, as it will trigger exceptions.