OtrInputStream: Decreased unconditional limit to 65536.
Decreased limit to 65536, which seems to be large enough in almost all
cases. Sanity check is still in place, so if enough data is available,
then lengths of 65536 or larger are fine too. A better sanity check may
be necessary.
Maybe we can verify if all parts of the 32-bit value are interpretable
as ASCII characters. In any case, this should not lead to memory
problems while still allowing for large amounts of data if required.