Clean up of unused field in SM.
The field that records whether or not a question is received, is never
read. The data is written to this field, but any other otr4j code
assumes that the question was successfully passed on to the
OtrEngineHost immediately after receiving the message.
With the removed field, a parameter becomes obsolete. This parameter is
also removed and any calls to the method modified accordingly.