Document that OtrExceptions on OtrEngineHost methods are deprecated.
The OtrException declaration for most of the OtrEngineHost methods does
not make sense for the simple reason that it interrupts the message
handling process for the protocol at unexpected times. These callbacks
are there for the sole purpose of informing the host application that
some event has happened. Any exceptions that are thrown are considered
as a faulty OtrEngineHost implementation.