Moved SM step 1 through 5 computations to State derivative classes.
Migrating code into the state implementation classes that are derived
from abstract class State. For traceability of the changes, I have so
far left the actual computations unchanged. Therefore, the code does not
currently compile. The method bodies however should be identical to
the method bodies that are remove from the original step1-step5 methods.

Next up is to modify the implementations such that the code actually
compiles and we leverage the State pattern's state transitions.