Moved proof.. and check.. methods to abstract State class for common use.
Moved proof.. and check.. methods that are heavily used in the
verification process for each SMP message, to the abstract class State
for common use. These methods are only of importance during the process.

Now that these methods have moved, we can also simplify these methods:
* SecureRandom instance available by default.
* G1 directly available as a constant.

For now I have left checkExpon and checkGroupElem as utility methods on
SM, although it may make more sense to move those as well.