Introduced final state CHEATED in case of an interruption of the normal process.
Introduced a new state "CHEATED" in case of an interruption of the
normal process. The state gets reset to default with the status of
CHEATED.
Interruptions are detected inside SM, i.e. outside the state machine
itself, and any exception except for SMAbortedException will be
considered cheating due to the fact that processing got interrupted
prior to completion.