Implemented 'approved' flag.
Implemented 'approved' flag indicating whether or not SMP exchange was
initiated by local user or verification was performed using shared
secret. (I.e. as opposed to using a question.)
I do not fully understand how this is inferior as the secret answer to
the posed question is still using the mechanism of a shared secret.