Remove 'approved' field in verification.
With the introduction of the otr v3 implementation a field 'approved'
was introduced. This field is supposed to indicate whether the
authentication initiative is approved by the user, i.e. this user - the
user of the client. However, the otr v3 spec documents nothing about
this. Furthermore, the distinction is based on whether or not
authentication is initiated by this user or whether no question was
posed.
(approved == authentication-initiated-by-user|| question-is-empty)

This flag does not make sense, because in both situations, the
underlying answer verification uses the exact same mechanism. The only
thing this 'approved' flag would indicate is that you don't trust the
counter-party to pose a smart question. Examples of dumb questions would
be questions that literally contain the answer or questions that suggest
the answer such that guessing becomes trivial. However, in that case
user should have denied authentication in total as there is no advantage
to establishing verification if counter-party behaves suspiciously.

All in all, the indicator should be used by the client and there it is
not clear how this flag should be used or what it means.