Fix v3 whitespace tag detection.
Logic accidentally tried to detect v3 whitespace tag by looking for the
v2 whitespace tag. That is, we check for the 3rd regex group which
contains the v2 whitespace tag. This was easy to confirm, as the
previous if-statement condition used the same group id. This is fixed
and code is simplified by removing the "group" notation around the
whitespace prefix. Now regex group 1 = otr v1, group 2 = otr v2,
group 3 = otr v3.
Added a number of tests to verify functionality surrounding whitespace
tag recognition.