Restructured transformSending in plaintext state.
Restructured transformSending for better state management.
Fixes bug where we possibly send a whitespace tag with no versions. otr spec documents that we should always send at least one version tag.
Verify that we have a viable policy (at least one OTR version available) before we start constructing the whitespace tag. If not viable, throw an IllegalStateException. This is considered a programming error as we should have prevented going into this branch.
Only set OfferStatus.sent if we finished serializing the message such that we can know for sure that we have constructed an offer. This is necessary as we only offer whitespace tags once. We do not want our offer to be with 0 possible OTR versions such that we can guarantee that the counter party cannot act on the offer.
Added unit tests for testing the basic characteristics of StatePlaintext.transformSending, such as whitespace tag sending, original message sending, etc. based on different policy configurations.
Added FIXME for later work to look at constructing an intelligent OtrPolicy object that will prevent us from going into branches that do not make sense for the policy configuration.