Fix REQUIRE_ENCRYPTION policy.
getRequireEncryption() would return true only if policy flag is set AND ALL protocol versions were enabled. This is strange as we would like it to return true even if no protocol version is supported. This policy defines how OTR should respond in case an encrypted session is not yet established, so we don't want it to respond too relaxed in some situations. If the policy is to *require* encryption, then under all circumstances is should behave as requiring encryption.
For this bug we already added a test in the previous commit.