OtrPolicy: deprecated policy configuration methods that allow you to pass in a value parameter.
It does not make sense to pass in 'false' to these methods as these methods modify the composition of basic policy values. How would passing in 'false' respond, for example in setEnableAlways. It does not really make sense to completely disable OTR. Instead, it should only respond to "true". Now the methods with 'value' parameter are deprecated in favor of methods without a parameter. This avoids confusion on its behavior and avoids accidentally disabling more than expected/intended.