Significant restructuring to introduce AKE state machine (AuthContext).
There is significant refactoring and restructuring in order to introduce state machine for AKE states.
This should be the last of the processes to be restructured and refactored into State pattern.
The effort is currently still in progress. Code does compile but tests do not yet run successfully and there are lots of FIXMEs to resolve before we can expect this code to run well.

Further improvements have been made to:
* support an OOP code structure and as a consequence to help avoiding mistakes,
* to create appropriate abstractions, such as:
  - the introduction of SharedSecret in crypto,
  - the introduction of the 'ake' package for the AKE process,
  - a strong reduction of LOC in AuthContext in favor of logic distributed over various ake.State implementations,
  - etc.

This commit message doesn't do it justice, however it's too late and there's too much to describe in order to capture it all. Unfortunately these changes are too massive to ever review although I knew that already going into this. However, it should make it easier to review the current code base, once the modifications are completed. There are stricter boundaries, nicer abstractions and lots of code with reduced access.