Removed AKEException with Context.InteractionFailedException.
We noticed before that the AKEException is an unfortunately chosen
exception name. The AKEException is not used within the ake package.
Instead, we use AKEException to signal that some interaction with AKE's
context failed. Now this is renamed to InteractionFailedException and it
is defined on the Context interface, as this makes more sense.
The primary function of InteractionFailedException is as part of the
public API in order to interact with AKE in exceptional cases.