Removed number of TODOs.
TODOs w.r.t. reducing duplicate code are removed for now as the repeated
parts are not captured in a method for the reason that we need to
preserve a reference to the produced data for later use.
TODOs w.r.t. using try-finally to ensure that we transition back to
StateInitial (NONE) are removed because right now we respond with a
reply message only if transition was successful. Try-finally does not
really add more safety for this situation as we already break processing
when an exception occurs. We would need to redo AKE anyways.