In case of receiving D-H Commit message in state Awaiting D-H Key: generate new key if swapping roles.
The otr spec describes the Awaiting D-H key state as tricky because you
may receive a D-H Commit message and in this case both parties are
assuming the same initiator role. To break this symmetry we compare
hashed public keys and respond based on its outcome.
In case ours is the lower value, we assume responder role. This change
now generates a NEW DH keypair instead of reusing the earlier generated
DH keypair.