Combined handling of AKE messages into a single private method.
Originally we had a separate handling method for each of the AKE message types. This is a heritage from the original implementation that had message processing for a larger part in the AuthContext class. Now that we have shifted the separation of concerns to a different (more sensible, hopefully) distribution, we can combine all the AKE message handling. We need to be careful though, for when to allow a receiver instance tag of 0. There are some FIXMEs to fine-tune that behavior, although this is mostly improving the structure such that it is "nicer". It should already be semantically correct.
Additionally, there is still a FIXME on checking message's protocol version against the expected protocol version. Although this is semantically not checked, we do already have a precaution inside the AuthState implementations such that we do not allow processing of mismatching versions.