Do not allow OtrEngineHost to throw OtrExceptions.
Do not allow or expect implementor to throw OtrException. This does not make sense as we expect the operations simply to succeed. In case we do need to handle for failure, we should do this in a way that also handles unchecked exceptions. And even so, we should define a different type of exception for it such that we can distinguish otr4j internal errors from "expected" exceptions thrown by the host implementor.