Test required cryptographic primitives in static initializer.
As follow-up to an earlier conversion to using runtime exceptions for
failures that should happen in case of missing support in JVM/crypto
libraries. We now want to ensure as soon as possible that all required
crypto types are available for use. Given that JCA works with factories,
we use a static initializer to request all required types from the
various factories. In case a 'NoSuchAlgorithmException' is thrown,
we now know that otr4j will not function correctly.