Modified getProtocolVersion to return version for current message state.
On top of that there's a number of javadoc additions. Further
investigation is needed in various methods that use getProtocolVersion()
to determine whether or not to OTRv3 and/or slave sessions are in use.
This does not seem to make sense, as the protocol version was originally
updated with practically every encoded message that was received. I
suspect that mixed OTRv2 and OTRv3 sessions would have made the protocol
implementation totally break down. This should not be necessary. As such
a lot of FIXMEs are added in order to further investigate current
findings.