Declined using ByteBuffer for reading message logic.
Given that we cannot really trust the data inside the messages that we
receive from the counter party, we cannot use the ByteBuffer. ByteBuffer
would throw runtime exceptions in case data mismatches with request,
e.g. requested number of bytes to read. By keeping to use InputStream we
can leverage the IOException that is possibly thrown to capture and
handle bad data in a sensible way.