Separate SMP concerns from other OTR concerns in OtrEngineHost.
Some methods were introduced only for the purpose of SMP. Some other
methods are introduced for calling back into the host in general, but
shouldn't be used in other parts of the code. Now we have defined an
interface specifically for the facilities that are required for SMP.
Only for those facilities should we call back to the host directly.