Removed forced replication of master AKE state to all slaves, independent of whether it is actually used.
The previous implementation of AKE support and AKE message handling
required forced updating of AKE states of all slave sessions to the
current AKE state of the master session. This results in possible ad-hoc
interrupted AKE negotiations and overruled the otherwise specified
protocol handling behavior in case of receiving certain messages. With
the new approach this is not necessary anymore. Now we replicate the
master AKE state to a specific slave session only for a single use case.
In all other cases we rely on protocol's AKE message handling behavior.