Improved work-around to switching outgoing session to new ENCRYPTED session.
Improved the work-around by implementing the switch evaluation in a
try-finally clause. In case we have actually done a transition and
the current outgoing session is a plaintext session, then we switch.
This way we can ensure that we at least progress into a better state.
The behavior is not completely the same as with the old implementation,
but at least it improves on the user's current outgoing session.
Leaving the FIXME for further evaluation.