Tweaks + resolved TODO: with new implementations of SessionKey and SharedSecret we no longer rely on sharing byte-arrays.
Only item left from original TODO is to avoid sharing TLV data through
original array instance. This is less of a problem given that it is only
for temporary use.
Additionally, defined TLV.EMPTY as 0-length byte-array for reuse in the
many cases where we want to send a TLV with 0-sized data payload.