Redesigned respondSmp() method to make more sense.
The original implementation did not make much sense, as it was deferring
execution to slave sessions, first based on instance tag, then based on
outgoing instance. Additionally, we defer from a method which receives
an instance tag to a method that does not accept an instance tag. So how
can we be sure that we actually handle for the requested instance?

With the new structure it is way more transparent how this is supposed
to working.