Moved SM protocol implementation in crypto to separate package 'session.smp'.
Extracted all static inner classes into their own (package-private) classes in package 'session.smp'. Most of the implementation is package-private as it concerns only internal state management. SM is the interface to the outside world and keeps a reference to the current state instance. The design of the logic has not changed, just the classes being extracted and SMP promoted to its own package.