Finished support for TLV 8 Extra symmetric key.
Finished the final piece for the extra symmetric key support. Now, whenever we discover TLV 8 in a message, we call back to the engine host to signal the discovery. The original message, the TLV data and the extra symmetric key are provided. The user message itself will also be returned as part of normal received message transformation.