Fix issue with shortened refreshSession mechanism.
In shortened refreshSession mechanism, we immediately send to receiver instance tag
with the D-H Commit message. This is possible because we already know it from the
currently established session. This prevents other clients connected on the same
account from also trying to initiate an OTR session. However, as we previously
copied the master session's authentication state to the slave session upon receiving
the DHKey message, we now need to initiate refreshSession on the master session
instead of the slave session. (More information in comments.)