Initial redesign of authentication state handling inside the Message State state machine.
The subclass underlying the Message State states implements the
AuthContext interface and supports storing AuthState instances.
This redesign is intended to simplify migrating the OTRv4 DAKE state
machine away from the separate Auth state machine implementation.

Currently both OTRv3 and OTRv4 Auth state machines are implemented as
the separated AuthState handling, but now migrated into the Message
state. Next step is to extract the OTRv4 states and implement them as
Message State types. Then we can evaluate whether or not we want to
allow having the (now OTRv3-only) Auth state machine running in parallel
with the Message states while being in OTRv4 message states (either one
of the DAKE states or the OTRv4 Encryped messaging state).

The use case of initiating AKE (OTRv3) or DAKE (OTRv4) with known
receiver instance tag is not especially supported yet. It may work, but
has not been evaluated. There is an open TODO for this.

Additionally, currently we need the integer type codes for (D)AKE
messages to be public as we use this to determine whether or not to copy
data from the master session state to the slave session state. This is
far from elegant, but it works. Conceptually everything should work, or
at least everything for which we have a test, but the current
implementation has its rough edges and needs further refinement and
refactoring.

As the moved and refactored logic has not been fully documented yet, the
build will fail.