WIP: implementing redesigned Double Ratchet initialization.
Currently implemented up to the point where it is almost usable, but not
quite. The current code is off on the j and k ratchet IDs, or better
said only one of the two. As a consequence, for the immediate start, you
might end up with messages that are sent for ratchet ID: -1, for which
the other party does not have decryption and MAC keys.

I've posed the question here: https://github.com/otrv4/otrv4/issues/192.
For now I'm not certain whether this is an implementation issue or a
spec issue. It seems to me that the current spec can indeed end up with
ratchet ID -1. To be continued.