Basic fuzzing for MessageProcessor based on JQF and afl.
Introduced basic fuzzing support using the JQF framework. JQF uses afl
for the actually fuzzing. It constructs a proxy for the Java-based code
and connects afl through it.

For simplicity, a Makefile is constructed that starts jqf-afl-fuzz in
the right way. It currently assumes necessary binaries are available.
JQF depends on afl-fuzz and its own binary jqf-afl-fuzz.

Various types of messages are provided as initial seed to the fuzzing
process. Seeds are collected in 'src/test/fuzzing-seeds'.