Introduce creation timestamp to determine whether to copy master session AuthState.
There is an outstanding decidability issue on whether or not to copy the
master session's AuthState instance when receiving the first AKE message
that contains the other party's instance tag. During initiation the
instance tag was not yet known, so we could only perform a AKE state
transition on the master session.
In a more exotic use case, when reestablishing a private messaging
session, we can send the initial AKE message directly with *the
known instance tag*. However, due to not being able to decide whether or
not to copy the state from the master session, this would cause issues.
Now we have introduced a timestamp, such that we can always assume that
the most recent of the AuthState instances is the leading instance.
Basic on this information we can decide to copy master session's
AuthState or continue with slave session's existing AuthState.