Describe how security policies can be used to protect crypto state.
Added remark about using security policies to prevent access to package
net.java.otr4j.crypto. This prevents base cryptographic secrets from
being accessed which are used to derive the working state of in-progress
encrypted "private message" sessions of OTR.

Originally, the intention was to prepare a ready-to-use or
active-by-default protection configuration, but Java does not really
allow preparing for this. This is mainly due to lack of guarantees when
the runtime context is not under control. Hence the prescription of how
to extend security when otr4j is used.