Include pgp-keys list with expected public keys.
It bugged me for a long time that the pgp-verify plug-in would just
download pgp keys as part of the verification process. If any plug-in
changed and was signed with a different key it wouldn't be detected
because the new public key would simply be assumed to be trustworthy.

I looked into checksum-maven-plugin as a way to detect if plug-ins are
according to expectations, but it is not well maintained.
pgpverify allows maven to set a list of expected pgp keys per
artifact, therefore it is possible to verify the keys of the downloaded
dependencies against the list of expected keys.