session: do not keep instance of ClientProfilePayload
Do not keep a local instance of ClientProfilePayload, instead request
OtrEngineHost for encoded client-profile payload every time the
client-profile is needed. This allows synchronization of all (master)
session instances by virtue of host-application managing the
client-profile. Both first construction and subsequent refreshes are
still managed within otr4j, by the session, but now the host-application
gets to decide whether and when to force a refresh, e.g. when policy or
keypairs change.

There are consequences to this, but we expect a (pro)active host-
applications to be aware of these. In passive applications, the host-
application would simply provide the keypairs, policy and client-profile
payload when requested, and otr4j sessions will manage the rest.