session: tweaked implementation, considering compare-and-swap
Tweaked the implementation for refreshing client-profile (payload).
Considering providing original/previous payload to make a compare-and-
swap possible, such that a host-application can decide for itself whether
atomically handling updates is necessary and if so have the necessary
information to do this conveniently. However, all of this seems a bit
excessive as it means that two sessions need to perform an AKE and also
refresh the payload in parallel. Currently, there is a risk of a profile
being generated that isn't persisted but instead is only used one time
for this one session, because the parallel refresh immediately overwrote
that generated profile. (This situation is annotated with REMARK.) This
would also have to happen with different contacts, because multiple
messages from a single account are currently handled sequentially, as
SessionImpl isn't thread-safe.